# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Golang Project siteAccess",
        "title": "SiteAccess",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Daniil56",
            "email": "daniil13.spiridonov@yandex.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/max": {
            "get": {
                "description": "gets the name of the site with the maximum access time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get name site with maximum time",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Site"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/min": {
            "get": {
                "description": "gets the name of the site with the minimum access time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get name site with minimal time",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Site"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/site": {
            "get": {
                "description": "gets the access time to the transferred site",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get the time on the site",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Example: yandex.ru",
                        "name": "site",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Answer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Answer": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "n/a"
                },
                "time": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "domain.Site": {
            "type": "object",
            "properties": {
                "site": {
                    "type": "string",
                    "example": "example.com"
                }
            }
        }
    }
}